Move test support class
