+++ tests/UserTest.php
@@ -3,15 +3,15 @@
 declare(strict_types=1);
 
 use LaraBug\LaraBug;
-use LaraBug\Concerns\Larabugable;
 
 use function Pest\Laravel\actingAs;
 
 use LaraBug\Tests\Mocks\LaraBugClient;
+use LaraBug\Tests\Support\CustomerUser;
 
 use function PHPUnit\Framework\assertSame;
 
-use Illuminate\Foundation\Auth\User as AuthUser;
+use LaraBug\Tests\Support\CustomerUserWithToLarabug;
 
 beforeEach(function () {
     $this->laraBug = new LaraBug($this->client = new LaraBugClient(
@@ -56,19 +56,3 @@ it('return_custom_user_with_to_larabug', function () {
 it('returns_nothing_for_ghost', function () {
     assertSame(null, $this->laraBug->getUser());
 });
-
-class CustomerUser extends AuthUser
-{
-    protected $guarded = [];
-}
-
-class CustomerUserWithToLarabug extends CustomerUser implements Larabugable
-{
-    public function toLarabug(): array
-    {
-        return [
-            'username' => $this->username,
-            'email' => $this->email,
-        ];
-    }
-}
